Implement questions and answers
Edit the data file containing the questions and answers. Separate the questions by sections by providing a name for each of them.

Change of the way of rendering of the questions and answers.

Implement functionality for calculating each question number - now we're precalculating cumulative question counts for each category and use these precalculated offsets to determine question numbers.