Replacing static paths with Vite imports in order to fix issue with the images in the URL after the deployment
Now we dynamically import images and PDFs for GH Pages compatibility.

Replace dynamic Tailwind bg URLs with inline styles for production.

Use imported assets for background images instead of static paths.

Import PDFs as modules for correct deployment paths.

Only the carousel background image is imported in different way (for that goal it was be created in the public folder, too).